immense cleanup; replaced improper uses of unsigned int with size_t
git-svn-id: https://paroxysm.googlecode.com/svn/trunk@440 1a4693bc-ef48-11de-b206-85b67c77eac7

+++ source/Container.cpp
@@ -23,7 +23,7 @@ Container::Container()
 
 Container::~Container()
 {
-    for (unsigned int i = 0; i < mWidgets.size(); ++i)
+    for (size_t i = 0; i < mWidgets.size(); ++i)
     {
         delete mWidgets[i];
         mWidgets[i] = NULL;
@@ -32,7 +32,7 @@ Container::~Container()
 
 void Container::display()
 {
-    for (unsigned int i = 0; i < mWidgets.size(); ++i)
+    for (size_t i = 0; i < mWidgets.size(); ++i)
     {
         if (mWidgets[i]->isVisible()) mWidgets[i]->display();
     }
